Prevent attempt at reset on connect
Soft-resetting the badge makes sense when dealing with non-native USB, but as we are connected over an CDC-ACM tty that's instantiated after ESP boot, we cannot reset without losing the port. In fact, the TiDAL does not implement the reset hint over USB DTS/RTS lines, so all this does is disable the port with flow control.